+++ src/app/Top3Posts.tsx
@@ -6,6 +6,7 @@ import { useSuspenseQuery } from '@tanstack/react-query';
 import { BoardRankResponse } from '@/types/board';
 import { useRouter } from 'next/navigation';
 import { Suspense } from 'react';
+import Top3PostSkeleton from '@/components/Skeletons/Top3PostSkeleton';
 
 function Top3Posts({ selectedStock }: { selectedStock: BoardRankResponse }) {
     const router = useRouter();
@@ -34,25 +35,6 @@ function Top3Posts({ selectedStock }: { selectedStock: BoardRankResponse }) {
                     </PostThumbnailCard.Footer>
                 </PostThumbnailCard>
             ))}
-        </div>
-    );
-}
-
-function Top3PostSkeleton() {
-    return (
-        <div className='flex flex-col justify-between relative border border-solid border-moneed-gray-5 rounded-[1.8rem] mb-[1.6rem] cursor-pointer py-[2.4rem] px-[3.2rem]'>
-            <div className='mb-[1.2rem] sm:mb-[1.6rem]'>
-                <h3 className='h-[2.8rem] mb-[.6rem] bg-gray-200 rounded-md animate-pulse w-[10rem]'></h3>
-                <div className='w-full h-[1.8em] bg-gray-200 rounded-md animate-pulse mt-[.8rem]'></div>
-                <div className='w-full h-[1.8rem] bg-gray-200 rounded-md animate-pulse mt-[.8rem]'></div>
-                <div className='w-2/3 h-[1.8rem] bg-gray-200 rounded-md animate-pulse mt-[.8rem]'></div>
-            </div>
-            <div className='flex items-center gap-[.8rem]'>
-                <div className='flex items-center gap-[.6rem] flex-1'>
-                    <div className='w-[2.4rem] h-[2.4rem] bg-gray-200 rounded-full animate-pulse'></div>
-                    <div className='w-[10rem] h-[1.6rem] bg-gray-200 rounded-md animate-pulse'></div>
-                </div>
-            </div>
         </div>
     );
 }
