+++ src/components/Chip.tsx
@@ -54,11 +54,3 @@ export const ChipLink = ({ className, active, label, href, disabled, children }:
         </Link>
     );
 };
-
-export const ChipSkeleton = () => {
-    return (
-        <div className='rounded-[1.2rem] px-[2.4rem] py-[1.2rem] bg-moneed-gray-4 animate-pulse'>
-            <div className='h-[1.4rem] w-16'></div>
-        </div>
-    );
-};
