+++ src/components/Skeletons/ChipSkeleton.tsx
@@ -1,0 +1,7 @@
+export default function ChipSkeleton() {
+    return (
+        <div className='rounded-[1.2rem] px-[2.4rem] py-[1.2rem] bg-moneed-gray-4 animate-pulse'>
+            <div className='h-[1.4rem] w-16'></div>
+        </div>
+    );
+}
