+++ src/components/Skeletons/StockRankButtonSkeleton.tsx
@@ -1,0 +1,7 @@
+export default function StockRankButtonSkeleton() {
+    return (
+        <div
+            className={`rounded-[1.2rem] px-[1.6rem] py-[.7rem] flex items-center text-[1.4rem] font-semibold bg-moneed-gray-4 animate-pulse`}
+        />
+    );
+}
